fixed ceil in MPCBAnalyzerNodesDecomposition::GetAcceleration

+++ magistracy/term_2/karpenko/lab2/mpcb_analyzer.h
@@ -66,7 +66,7 @@ class MPCBAnalyzerNodesDecomposition: public MPCBAnlyzer {
     for (const auto &stat: stats)
       nodesToCompute += stat.nodesComputable;
 
-    int nodesToComputePerProcessor = ceil(nodesToCompute / stats.size());
+    int nodesToComputePerProcessor = ceil((double)(nodesToCompute) / stats.size());
     double solutionTimeParallel = 2 * T_S + (DIM + M) * nodesToComputePerProcessor \
         * L * networkDiameter * T_C + T * nodesToComputePerProcessor * computionComplexity;
     double solutionTimeSequential = T * nodesToCompute * computionComplexity;
